+++ package.json
@@ -4,10 +4,10 @@
   "description": "",
   "main": "src/index.js",
   "scripts": {
-    "start": "npm run build && node ./build/app.js",
-    "start:dev": "nodemon --exec babel-node ./app.js",
+    "start": "npm run build && node ./build/index.js",
+    "start:dev": "nodemon --exec babel-node ./src/index.js",
     "clean": "rm -rf build && mkdir build",
-    "build-babel": "babel -d ./build ./app.js -s",
+    "build-babel": "babel -d ./build ./index.js -s",
     "build": "npm run clean && npm run build-babel",
     "test": "mocha -r esm --exit ./src/tests/*"
   },
@@ -22,18 +22,11 @@
   },
   "homepage": "https://github.com/waterloop/autodoc#readme",
   "dependencies": {
-    "@babel/cli": "^7.14.8",
-    "@babel/core": "^7.14.8",
     "chai": "^4.3.4",
     "esm": "^3.2.25",
     "mocha": "^9.0.2"
   },
   "devDependencies": {
-    "@babel/node": "^7.14.7",
-    "@babel/plugin-transform-runtime": "^7.14.5",
-    "@babel/polyfill": "^7.12.1",
-    "@babel/preset-env": "^7.14.8",
-    "@babel/runtime": "^7.14.8",
     "nodemon": "^2.0.12"
   }
 }
